requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pdfplumber==0.11.4
openai==1.35.7
langchain-core==0.2.23
chainlit==1.1.202
requests==2.31.0
